Added control to change h1 etc to the hier name.
When there is more than one column in the data frame going into the
function pop_splitter, it will append new columns onto the data frame.
This will make the data frame bigger by the number of hierarchies
within the first column of the data frame. The new addition checks for
this size and makes sure that the hierarchy designation is the same and
will rename those columns based on the submitted hierarchy.