simplified bruvo.dist and bruvo.boot
By piecing off the preparation of data into an S4 object and the
calculation of Bruvo’s distance into a function, much of the bruvo.dist
function could be replaced with three lines.

The new form of the S4 object, originally for bootstrapping purposes
runs faster than the original implementation.