Further work on bitwise_distance.c
bitwise.dist is mostly complete now.
Missing data is currently unhandled, and defaults to matching missing points with other missing points or natural 0's
If no 2's are present in a sample, it will treat it as a haploid and the data set will be rejected by the wrapper function bitwise.dist()
Unit tests have not been created yet.
Aside from these points, bitwise.dist() is functioning as intended.