bitwise.pgen has been added for calculating the pgen values of individuals in a genlight object.
pgen represents the probability of an MLG being produced via random mating.
This function returns, by default, the logarithm of the actual percentage value in order to increase the range of outputs.
An option is available to return the actual percentage values, but these values will typically round to 0.
Function tests should be added for this function in the future.