Added new features to bruvo.msn and poppr.msn
The user can now specify include.ties=TRUE when calling bruvo.msn or poppr.msn
If selected, the resulting msn will include any edges that could have been included in the msn.
This feature is off by default.
Unit tests for this feature will be included in a future commit.