modified MLG methods
I have removed the setmlg method. This method should be for genclone objects as the MLG objects should not be visible to the user.

I changed duplicated.MLG and unique.MLG to duplicated and unique. I did the same as I did with "print" where I set a generic and then defined the method. These are hidden and they act much like the base function.

I added methods for levels so that the custom MLG definition can be easily changed.