improved new.poppr.plot. Add method for ialist class.
the plotting function will now display the numeric value of the observed data and the pvalue to three sigfigs. This is constructed in such a way that it will always display the number of permutations utilized, but will optionally display the number of samples, the population name, and the data name. This has been implemented as the default plot method for ialist (undocumented)