add MLG, mlg.custom, and mlg.levels functions.
MLG is an analog to mlg.vector, but it will give the user access to all levels of MLGs

mlg.custom is specific for setting and getting custom mlgs

mlg.levels will allow the users to set the levels of the custom mlgs. This is useful for situations where the user wants to replace spelling errors.