Add normal def. for CI, still problematic
I'm using boot.ci's normal definition for confidence interval, which corrects bias, but does so in a way that has the side effect of potentially pushing the confidence interval the opposite way. I could force the confidence interval to be centered around the observed value, but that seems like it's cheating.