Address weird testing bug
Test that would pass, but then would fail due to an
error in the summary:

Error in vapply(test_results, "[[", TRUE, "passed") :
  values must be length 1,
 but FUN(X[[5]]) result is length 10

This was due to a call to `testthat:::sumarize_one_test_results()`
could not handle a single results element in the results list containing more than one result. I should probably make a MWE and submit it as a bug.