fix default distname for MLG objects.
This additionally fixes the default distanme for
genclone objects. The reason for this change is
because it was realized that using "nei.dist" didn't
make much sense as a default distance for collapsing
MLLs. In #76, the arguments for mlg.filter were
changed so that "diss.dist" was favored for genclone
objects, but what slipped past my radar was the fact
that if no distance was specified, it was taken from
the distance measure specified in the object itself,
which was still initializing as "nei.dist".