Fix plot of `genotype_curve()`: x-axis continuous
The genotype curve initially had set the x axis to
have a grouping on the factor of the number of loci
to appease the `geom_boxplot()` gods. Moving the
grouping to `geom_boxplot()` allowed me to free up
the x axis to be a continuous variable, meaning that
the user can now add `geom_smooth()` to the plot
without much trouble at all! This was brought to my
attention by Kara Woo's response to a question from
Stats for bios on Twitter:

https://twitter.com/statsforbios/status/783334647249993733
https://twitter.com/kara_woo/status/783336540407685120