add catch to prevent fencepost error
This will fix #169

Because the while loop was accessing the array of missing data
after the counter was incremented, a block from outside of the
array was being accessed. I attempted to move the allocation to
the beginning of the while loop, but was having problems, so I
added a control structure instead. It's not the best solution,
but it will work for now.