remove tilde
I'm doing this, not because it is the correct thing to do, but because I
do not want to have to figure out what the hell CRAN is complaining
about for a citation that has existed on CRAN for several years.

It sucks because ñ is NOT the same as n, but unfortunately, CRAN
complains when we include non-ASCII characters in our code.